feat: change the voting model based on the proposal action
